Edited the Higher Level APIs
Removed ‘specific’ from the function names
Made the functionalities and flags consistent to only return 1= success for the particular operation and 0 = failure
Replace ‘sleep_time’  ‘retry_delay’
Comment the low level api instead of C function calling
6) Line 158: "print *, "Flag is not set for sending, retrying…””
What does it mean? Why not set for sending? Please give a more meaningful print.
print *, "Flag for the variable you are sending is on, so wait for the other client to receive it"
Suggestions over it to change, since we might have risk of overwriting
Removed status variable from the ‘retrieve_variable_size’, because it’s redundant.